Rebuilt Sampling Patterns and fixed VRAM graph
I decided to rebuild the graphics for the sampling patterns in R. This made it easier to place points at the desired coordinates and to maintain consistent characteristics.
Also updated the Means + VRAM graph, moving the legend over. (The script already did this but the PDF used was out of date.)