Added handout version
I have fixed the framing for the presentation so the beamer packages's handout option will produce a useful output. This did involve a very minor change to the Presentation.tex file
The Handout.tex file is identical to the Presentation.tex file except for the lines necessary to generate a handout version of the presentation.